Update mealMaker.js
Create an if…else statement to check if _meal and _price values exist (or are truthy values). If so, return a string telling potential website visitors what Today’s Special is. If _meal and _price values do not exist (or are falsy) return the string 'Meal or price was not set correctly!'